MediaPlayer: detect and convert encoding in subtitles.
The SRT file format unfortunately does not specify a text encoding (in
fact, there isn't even really a specification for it). Just loading the
bytes and assuming UTF-8 does not work so well, so we now try to detect
the encoding from the first line of text found.

Fixes #9668

(sometimes watching a movie is all an adventure!)