DeskBar: remove app name from expanded window items
Expanded window items are listed right under the team item. So, it is
not needed (and a waste of space) to repeat the team name there.

Detect two common patterns for adding the application name to windows:

AppName: Window

and

Window - AppName

Remove both from the item labels, leaving a little more space for useful
information in DeskBar.