Webpositive: unset tool tip over tab close button
Tool tip reads tab name or "New tab" for new tab, making it sometimes
appear that the close button would produce a new tab instead. Removing
the tooltip over the close button eliminates this confusion.

An alternative suggested in hrev45298 was to make the tool tip say
"Close tab" instead but some felt it unneccessary and so was reverted.

"A tool-tip for the close buttons is completely unnecessary. The
prevailing thought for items where the action is easy to grasp (or easy to
grasp upon clicking the first time), is that a tool tip is not necessary."

This method fixes the confusion w/o adding uneccessary clutter.