Add libmedia.so to the minimum image
libmedia.so was used by at least 2 apps included in the minimum image,
Activity Monitor and Deskcalc which subsequently failed to load on minimum.

Deskcalc wasn't actually using libmedia.so so I managed to get it to link and
run on minimum, however, Activity Monitor is using libmedia.so so needed it
to be present.

To fix this I added libmedia.so to the minimum image, I hope this does not blow
its size budget. (adds 1.0MB)