Workspaces: Make sure settings is set only once
... fixing oversight in hrev51512.

Erase the file and write all settings to disk each time.

When Workspaces is running in a window WorkspacesSettings
is destroyed first, then WorkspacesView is destroyed.
WorkspacesSettings saves all settings except for 'switch on
wheel', then WorkspacesView reads the settings in from
disk, tacks the switch on wheel setting on, then writes
the settings back to disk.

If WorkspacesView is a (Deskbar) replicant instead the
WorkspacesSettings steps are skipped. WorkspacesView reads
in the settings from disk, tacks 'switch on wheel' on, erases the
file, then saves all settings back to disk. Use ReplaceBool to
ensure that the 'switch on wheel' settings is added just once.

Create the file if it doesn't exist in either case.

Don't check Unflatten status because will fail on empty file.

SaveSettings on the view before opening Deskbar replicant
so that 'switch on wheel' gets loaded.