De-wire net_stack_cleanup from boot/loader/main.cpp
This reverts commit 2960780faa7e9c2e5e7a80cb8925be5ee59ae5ef.
Sorry for the noise; I thought I'd run a compile but apparently it was
only a partial one.

It seems that the netboot core is linked into the main bootloader
no matter the build configuration, which means that since this one
function was called, the linker tried to resolve all the other symbols
from that static library. Which of course failed.