HaikuDepot: Handling Corrupted Download Data
Damage in latter portions of the downloaded package data can cause the
meta-data for the caching to load OK, but the actual data is unable to
be parsed.  In this case, the application can fail to refresh the
broken data.  This change will mean that the data is "moved aside" into
another file so that on the next run, the application will load fresh
data in place.