Fix first replicant wrapping bug on wide clock
In this commit:
a44504a168548e129e58eeace336a6a2315257ad
    Deskbar: Refactor TRelicantTray::LocationForReplicant()

I made the following assumption that:
    * if index == 0 return right away, no calculation required.

But this assumes that the first replicant will always fit on the
same line as the clock which doesn't work if you have a wide clock.

Update to always calculate the position even of the first replicant
in vertical (multi-row) mode.

In horizontal mode we still skip some calculations for the index 0
case.

Also I was using kClockMargin (12px) for the right margin, the
right margin should be only 8px: 5px for the dragger plus 3px more
for kTrayPadding. Finally we add in the width of the clock and
extra clock margin iif we're on first row and clock is not hidden.

Fixes (hopefully the last of) #8641