docs/misc: Add BeIDE_ProjFileFormat.
This is the beginning of a large "move developer docs from the wiki
to the tree" operation, which will probably take some time to complete.

The general goal is to consolidate all docs that would be used by developers
(i.e., anyone working on the Haiku tree) into the tree itself. Docs on
getting started contributing, or for translators, designers, etc. will remain
on Trac and on the website.